Backend/requirements.txt
Streamlit>=1.29    # Frontend framework for web apps
Supabase>=2.0.2    # Supabase client for database operations
FastAPI>=0.104.1    # Backend API framework
Uvicorn>=0.24.0    # ASGI server for FastAPI
Python-dotenv>=1.0.0    # Environment variable management